# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuickLink API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/webhooks": {
      "get": {
        "operationId": "listWebhooks",
        "responses": { "200": { "description": "Webhook endpoints", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEndpoint" } } } } } }
      },
      "post": {
        "operationId": "createWebhook",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWebhookInput" } } } },
        "responses": { "201": { "description": "Created webhook endpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookEndpointWithSecret" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WebhookEndpoint": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "url": { "type": "string" }, "isActive": { "type": "boolean" },
          "events": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string" }, "updatedAt": { "type": "string" },
          "deliveriesCount": { "type": "integer" }, "secretHint": { "type": "string" }
        },
        "required": ["id","url","isActive","events","createdAt","updatedAt"]
      },
      "WebhookEndpointWithSecret": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "url": { "type": "string" }, "isActive": { "type": "boolean" },
          "events": { "type": "array", "items": { "type": "string" } },
          "secret": { "type": "string" }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" }
        },
        "required": ["id","url","secret"]
      },
      "CreateWebhookInput": {
        "type": "object",
        "properties": {
          "url": { "type": "string" },
          "events": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["url"]
      },
...
    }
  }
}